Fix freetype unbundling and omit pdf_bundle_freetype=false
There are three issues here relating to freetype: an upstream
unbundling bug, a bug in the reversion of unbundling and the
use of an unnecessary build flag.

We need to add an upstream patch from the main branch to fix
build errors related to unbundling freetype.

The bug in reversion of unbundling relates to a need to run
our unbundling find command through build/config/freetype in
addition to the base/third_party and third_party directories.
If we don't then the original file will not get restored.

The pdf_bundle_freetype flag is set to false by default which
means that its unbundling status is determined by the value
of use_system_freetype in build/config/freetype/freetype.gni.

(Note: attempting to build with pdf_bundle_freetype=true
together with bundled freetype gives duplicate symbol build
errors; ie do not try to force bundled freetype for pdfium).

We might as well omit any usage of use_system_freetype and
leave this to the unbundling process, which replaces the
flag via replacing build/config/freetype/freetype.gni with
the unbundled copy (build/linux/unbundle/freetype.gn).