Build with enable_screen_ai_service=false
This is some kind of new AI-related service used by the
accessibility component.

There isn't much information available on this, but from having
a quick browse of the source code it seems like it uses OCR and
might have some kind of screen-reading capability.

It could possibly be related to the new AI search capability
which was covered by the tech press a few months ago [1] [2].

Let's err on the side of caution and disable this.

[1] https://www.androidcentral.com/apps-software/google-search-your-screen
[2] https://blog.google/products/search/visual-search-ai/