Add back previously removed convertutf.patch
This was removed a long time ago during the period when
upstream debian chromium was effectively unmaintained and
the patch would no longer apply cleanly.

The idea behind the patch was to use libicu to replace the
embedded utf conversion in the breakpad component (this was
actually done for licensing reasons, with efficiency being
a neat side effect). As the patch is now maintained, the
reasons for dropping it no longer apply,

For those not using system icu I have included a version
of the patch modified to use the bundled icu header
instead pf the system one.