configure.sh: Re-organise and add more section headings
This commit moves a few items around so that things are
ordered, grouped together, labelled and better described.

Also remove a couple of obsolete comments at the start of
the Clang/Polly configuration section.

There shouldn't be any functional changes, and the whole
script ought to be a lot easier to follow.