Let the MARCH quirks handle things if MTUNE is set to generic
There's no need to skip setting MARCH to generic (via the use
of a conditional) if MTUNE has been set to generic. Just let
MARCH be set to generic and have the quirks section immediately
below catch this and force MARCH to x86-64-v2.

This commit is just a simplification with no functional change.