configure.sh: Replace some grep tests with sed ones
Using grep inside a test, eg [ "$(grep foo file)" ], can
cause a shell script to exit, since grep will exit with
an error if the string it is searching for is not present.

Although this doesn't appear to be an issue in our script,
we can avoid it entirely by converting each grep to a sed
equivalent.