Simplify/fix aggregation of deps_disable/deps_enable
In general, we only need to match the first part of
libfoo-dev, eg by specifying libfoo, as long as this
produces a unique match.

If this for some reason doesn't produce a unique
match, eg there are both libfoo-dev and libfoo-bar-dev
dependencies, then we must take care to specify
'libfoo-dev' instead of 'libfoo' if we don't want
to also enable/disable libfoo-bar-dev.

Luckily, I don't think we've actually been harmed by
the previous way of doing this.