Re-work configuration of printing
Upstream printing code has changed, and we are forced to
disable the new out-of-process printing method.

There is no X11-specific OOP print dialogue anymore, and
in order to build without print preview (which crashes)
we have to disable the new OOP stuff.

The OOP_PR variable is dropped and we add PRINT_PREVIEW.
The latter will build with enable_print_preview=false and
enable_oop_printing=false when set to zero (the default).

The X11_ONLY variable remains, and now just becomes an
alias for OZONE_WAYLAND=0 when set to 1.