configure.sh: Add basic DoH configuration and policy handling
The DNS_BUILTIN variable allows for the enablement of the
built-in DNS-over-HTTPS (DoH) client in the policy file.

DNS_INTERCEPT allows for the disablement of DNS interception
checking. These checks are enabled by default, and help
protect against DNS hijacking/poinsoning, but might not
be wanted under certain circumstances (they create some
extra traffic upon DNS initialisation).

DNS_HOST allows one to change the DoH host resolver address
to something different to the default. Refer to the
configure-variables.txt file for more details.

By virtue of the above, this commit puts in place some basic
handling of changes to policies.json via use of the internal
POL variable.

We can remove the final hunk of the bromite DoH improvements
patch. The code under 'if (check_parental_controls)' won't get
executed due to the fact that we use a policy file to set
DnsOverHttpsMode to secure.

We leave the other hunk alone, even thought it might be not be
necessary, just to be on the safe side.