Fix/improve/simplify VERSION variable handling
First of all, there's no need for a nested case
statement, as we can check for invalid strings
under one such statement.

What's perhaps worse is that I somehow failed to
spot that we were (pointlessly) checking for a
blank string twice under the inner statement.

Also, if we are going to check for '-1' as an
invalid version string, then we should also
check for its stable counterpart '-stable1'.

Finally, use [1-9] instead of [0-9] to match
the revision number, since '-0' is never used
as a revision.