configure.sh: Re-work system/bundled OpenH264 support
We introduce the SYS_OPENH264 variable in order to
distinguish between system and bundled support, and
convert OPENH264 to function solely as a general
feature toggle (to permit any form of OpenH264 to
be disabled).

When NON_FREE=0 and OPENH264=1 (the latter is the
default), we ensure that SYS_OPENH264 is enabled
so that the free system library is used (ie to
safeguard against anyone configuring with both
NON_FREE=0 and SYS_OPENH264=0 and ending up
building the non-free bundled library).

When OPENH264 is set to zero, we now additionally
disable the rtc_use_h264 build flag (in addition
to media_use_openh264) to ensure that OpenH264
support is fully disabled.

Also make clear that SUPERVISED_USER=0 depends
an optional non-free bromite patch.