Always apply rtc-protobuf.patch
This patch sets rtc_enable_protobuf to false.
According to the blurb it's related to making
debug recordings, so seemingly debug-related.

Disabling this flag obviates the need to patch
third_party/webrtc/test/BUILD.gn in the optional
catapult patch, thus simplifying it.

We move optional/disable/rtc-protobuf.patch to
various/disable and allways apply it, not just
if disabling catapult.