configure.sh: Tidy up the template check section
We do the following :-

Drop PF_DIR and use P_FILE to point to the policy
file (without a debian directory prefix). This
simplifies things a fair bit.

Remove the unnecessary curly braces from the file
variable in the template check loop.

As we now only do one task within the loop, move
the comment outside the loop.