configure.sh: Stop aiming for repeatability
It's easier all round to aim at making the script
a one-shot affair, ie running it once and not
accomodating the ability to re-run it.

The only time the script is safe to re-run is
when it errors out before any files are touched
(eg when the LTO_DIR path does not exist).

Apart from 'early exit' exceptions, if you need
to re-run then delete the debian directory and
start afresh.