Use shell parameter expansion instead of dirname
Looping dirname 10,000 times in a test script
was something in the order of 100 times slower
than using the parameter expansion equivalent.

As this builtin command is only used a couple
of times near the top of the script, there's
no reason not to use it.

Add a small explanatory note to explain what
external command the builtin one is replacing.