Minor re-work of the crashpad handler patches
The third patch was originally intended to completely
stop any elements of the carashpad application from
being built. It isn't strictly needed, and could
possibly introduce subtle issues. For the time
being disable it.

Most of the second patch is just altering file paths
to prevent build errors. Move the two hunks which
aren't involved in this over to the first patch.