Minor re-work of the crashpad handler patches
The third patch was originally intended to completely
stop any elements of the carashpad application from
being built. It isn't strictly needed, and perhaps
could cause subtle problems. So disable it for the
moment.

Most of the second patch is just altering file paths
to prevent build errors. Move the two hunks which
aren't involved in this to the first first patch.