Fix up debian/shims/chromium.sh
We can get rid of repeated calls to the external expr
command by using a case statement to match the files
that we want to avoid sourcing.

We can exit early if running without gdb in debugging
mode, and thereby avoid executing 'mktemp -d'. Note
that doing this avoids leaving behind a stray
temporary profile directory.

We can simplify the use of the exec command to start
chromium by taking the non-exec version out of the
else branch of the 'if..then' test (nothing below
the exec statement will get executed).

We can also factor out the 'rm -rf $TEMP_PROFILE'
command and execute it just before the end if
want_temp_profile=1.

the 'exit $?' command needs to be moved to the end of
the script so that both branches of the want_debug
test can get a proper exit code.

We also make the spacing more consistent, replace
backticks with '$()', and where appropriate convert
some 'if..then' commands to case statements.

There should be practically no functional change.