Improve some of the build commands in d/rules
Printing the message about not re-doing unbundling
at the beginning of autoconfigure does not really
help us much, and just adds noise.

Replace the test for the non-existence of the
source file (out/Release/locales/en-US.pak) with
a test for the existence of the destination file
(out/Release/resources/en-US.pak). This way we
would error if the mv command failed due to
the source not existing, rather than having
what would be a significant error ignored.