Build with enable_maldoca=false
The maldoca component enables analysis of Office
documents for malicious macros.

Since the UC project disables things like safe
browsing, as well as potential for privacy issues
arising from the use of this component, we will
disable it.

Something like this is probably best dealt with
outside of the browser, eg using antivirus, not
automatically opening office documents, etc.