Allow skia to use bundled harfbuzz
As with the situation with freetype, skia forces a
dependency on system harfbuzz (both stable and unstable
use bundled harfbuzz).

We can patch this in a similar fashion by making the
internal skia_use_system_harfbuzz flag to depend on
use_system_harfbuzz.

We might as well amalgamate this change with the skia
freetype patch and simplify configuartion by applying
it unconditionally.

Move/rename :-

optional/fixes/skia-allow-bundled-freetype.patch

to...

various/fixes/skia-allow-bundled-freetype-harfbuzz.patch