Use SKIA_GAMMA to set a gamma in the range 1.0 to 3.0
Instead of just being a boolean to enable/disable the
corresponding patch, make the value of SKIA_GAMMA be
he actual gamma value.

The old behaviour is preserved whereby a value of 0
disables and 1 enables the patch. A vaule of 1 then
effectively becomes an alias for 2.2 (the value in
the unaltered patch).

Set SKIA_GAMMA=1.0 rather than SKIA_GAMMA=1 to
actually set the value (as opposed to enabling).

Otherwise set any value within the 1.0 to 3.0 range.