configure.sh: Do not enable INTEL_CET by default
Upstream support for this is only available for
windows builds via the enable_cet_shadow_stack
build flag.

Searching for this flag lead me to a guard placed
around code in v8/src/builtins/x64/builtins-x64.cc
which might possibly be needed for linux-based
CET too.

As I don't have a new enough CPU to test CET, or
know for sure whether the v8 patch is needed,
I've marked this variable as experimental and
disabled it by default.