Simplify usage of mktemp and non-debug program startup
As with the setting of tmpfile via mktemp, exit if
the setting of TEMP_PROFILE fails.

In the case of the former, rely on mktemp for any
error messages instead of suppressing them and
using our own non-informative version.

We can simplify the non-debug execution of $CHROMIUM
by using booleans to decide whether to use exec or
not (based on the value of want_temp).