shell launcher: Prefer gxmessage for showing errors
In the output_error function, we test for
gxmessage, then xmessage (part of x11-utils)
and fall back to the echo builtin.

We prefer gxmessage as many of its dependencies
are shared with the main ungoogled chromium
package or otherwise are likely to be installed
on most systems. It's perhaps a better option
for wayland users too.

We add gxmessage with x11-utils as an alternative
to the 'Recommends' field of debian/control.in.

Neither package is added as a hard dependency as
neither are now absolutely required (one can
always run the launcher from the terminal to
investigate errors).