Disable AlternateErrorPagesEnabled via managed policy
This is an alternative to a vandium patch [1], which
potentially stops a web service from being used to
handle navigation/connectivity problems. Using such
a service is probably not a good idea from a privacy
perspective.

Excerpt from the microsoft edge policy documentation [2] :-

"Allow Microsoft Edge to issue a connection to a web service
to generate URL and search suggestions for connectivity issues
such as DNS errors.

If you enable this policy, a web service is used to generate
url and search suggestions for network errors."

[1] https://github.com/GrapheneOS/Vanadium/blob/main/patches/0015-disable-navigation-error-correction-by-default.patch
[2] https://admx.help/?Category=EdgeChromium&Policy=Microsoft.Policies.Edge::AlternateErrorPagesEnabled