Fix configuration/building with WEBGPU
This wasn't finished from a previous commit which
enabled configuration of webgpu.

In 122, both the skia_use_dawn and use_dawn build
flags have been changed to be true by default. This
means that we need to build with skia_use_dawn=false
when WEBGPU=0 (the default) in order to avoid a FTBFS.

If WEBGPU=1 we need to enable the the runtime flag
in /etc/chromium.d/gpu-options.