Fix crashes and screwed-up interface
There were issues in the tab strip context menu and
zoom reset button in the last release, which were
probably behind some of the crashing.

After a week of debugging and tearing everything
apart, the cause was annoyingly simple. A previous
change meant that the locale pak file wasn't
getting copied and updated.

Similar changes to d/rules and d/configure.sh to
make things a 'one shot' affair were sound, except
for this one, which in hindsight I should have
left alone.