Lower the hot function multiplier from 15 to 12
The RFC for LLVM's Machine Function Splitter [1] mentions
aggressive inlining as being detrimental to performance.
With this in mind, we can perhaps afford to be somewhat
less aggressive with our hot function import limit.

Bringing this limit down to just under the upstream
default seems reasonable, going from 360 (24x15) to
288 (24x12). This compares with an upstream default of
300 (30x10).

This should result in a slightly lower binary size and
possibly provide a tiny performance boost.

[1] https://groups.google.com/g/llvm-dev/c/RUegaMg-iqc/m/wFAVxa6fCgAJ