Disable media remoting by default
Chromecast is crippled in upstream UC, so media
remoting can be safely disabled.

We can remove enable_media_remoting_rpc=false
from debian/rules, as this will automatically
become false when enable_media_remoting is set
to false.

There is an unmerged pull request which enables
chromecast [1], which I will implement in a
subsequent commit. However I will keep the media
remoting configuration option so that it can be
disabled independently of chromecast.

[1] https://github.com/ungoogled-software/ungoogled-chromium/pull/2116