configure.sh: Allow enabling chromecast
Introduce a CHROMECAST configuration variable which
when set to 1 enables chromecast WITHOUT enabling
and disabling the respective '--enable-mdns' and
'--media-router=0' runtime flags.

When set to 2 everything, include the runtime flags
(in /etc/chromium.d/network) will be enabled.

A number of changes need to be made in order to
enable chromecast. Firstly, we need a patch from
upstream pull request #2116.

We also need to take care of a vanadinite patch
which disables media router by default.

We make changes to the configuration script to
enable three build flags via deleting the following
build flags from the submodules's flags.gn file :-

enable_mdns=false
enable_remoting=false

We need enable_service_discovery to be false, but
its value mirrors that of enable_mdns (which is
true by default). We have been deleting the former
routinely for some time due to this fact.

The EnableMediaRouter managed policy key is deleted
so that we rely on the flags. It's easier to rely
on runtime flags since everything required to enable
chromecast is in the same place and annotated with
comments.

The media remoting components in configure.sh have
been changed such that disablement is no longer the
default.

Media remoting can however be disabled independently
of chromecast by configuring with MEDIA_REMOTING=0.

Move (and comment out) :-

vanadinite/0008-disable-media-router-by-default.patch
to
optional/disable/media-router.patch