configure.sh: Always use managed policies
There's no real use case for setting POLICIES=0, so
we can simplify things a bit by removing the POLICIES
variable.

The alternative is to have to put more and more things
within the if [ $POLICIES -eq 1 ] block. I wasn't doing
this with ZSTD, and realised it was easier to remove
this variable entirely.