configure.sh: Fix handling of CHROMECAST/MEDIA_REMOTING
This went through a few changes and I got two
of the CHROMECAST conditionals mixed up. The
first condition should have been an '-eq 0'
rather than and '-ge 0'.

Like with TRANSLATE and LENS, we allow setting
the second CHROMECAST conditional to a value of
2 or greater (to additionally enable any rewuired
runtime flags). This is so that say, mistyping
say 3 instead of 2 doesn't affect things.

We can also set MEDIA_REMOTING independently
of CHROMECAST at the top of the script.

Note: the chromecast deb release probably
wasn't affected.