Omit image writer utility/removable storage writer service
Image writer utility appears to be a chromebook-related
component and separate to the image writer private api.

The only thing linking the utility to the api seems to be
the removable storage writer service, which is a dependency
of the former and a header include in the latter.

This commit removes both the image writer utility and the
removable storage writer service by making both non-linux
options.

Added: various/disable/image-writer-utility.patch