Use the upstream default import limit
Benchmarks indicated that a function import limit of
somewhere between 30 and 50 is optimal.

Let's try going with the default of 30, which seems
to be around the lower end of the optimal figure, but
keep the hot multiplier of 12 (default is 10).

This give a hot function limit of 360 (30x12), which
equates to an import limit of 36 (36x10 = 360) when
using the default multiplier of 10.