Really fix SYS_RUST=1 (system packaged rust)
First of all, I erroneously commented out the deps_enable
line instead of the op_enable one, to disable rust.patch.

I also left in an old gn_enable on clang_version, which
is now set within the SYS_CLANG section.

The existing rust patch, which wasn't needed, can now
be replaced with a new one which solves a FTBFS error.

The problem was that older versions of rust (like that in
debian unstable)  lack support for the 'split-lto-unit'
build option, so we need to patch build/config/rust.gni
to make the internal 'toolchain_supports_rust_thin_lto'
build flag false.

Upstream debian chromium doesn't run into this issue
because they disable lto (and also cfi, which requires
lto). They did seem to indicate that the clang_* build
flags were needed, but this wasn't enough.

I'm not sure if toolchain_supports_rust_thin_lto is a
wise idea or not, so I have added a note to the
configuration-variables.txt documentation to
recommend SYS_RUST=2 instead.