Re-work/amalgamate the disable swiftshader patches
This is in preparation for enabling swiftshader by
default and allowing it to be disabled by the
configuration script.

Before doing this we amalgamate the two swiftshader
patches into various/fixes/swiftshader-build.patch.

We alter the declarations of three swiftshader build
flags to make them depend upon the value of
enable_swiftshader.

In other words, we make enable_swiftshader into a
master switch to enable/disable the following flags :-

angle_enable_swiftshader
dawn_use_swiftshader
enable_swiftshader_vulkan

Even if we build with swiftshader enabled, we can
still disable any of the individual flags above.

As an aside, enable_swiftshader_vulkan now gets
disabled when enable_swiftshader is false, which
wasn't the case before (it was missed out).

I have put build flag guards around the swiftshader
code which was disabled in the original patches.