Build WebGPU by default
First of all, rename the gpu-options flag file from
gpu-options to gpu, as the '-options' suffix isn't
really needed.

The configuration options for WEBGPU are extended
so that to fully enable out of the box one needs
to configure with WEBGPU=2 :-

0 - disable
1 - enable (need to edit /etc/chromium.d/gpu)
2 - enable completely

The default is 1. There are security concerns which
surround WebGPU, and users will need to enable the
'--enable-unsafe-webgpu' runtime flag to enable the
feature.

SWIFTSHADER_WEBGPU needs to default to 1 to accord
WebGPU being enabled by default. The SWIFTSHADER
configuration needs to be moved below WEBGPU due
to SWIFTSHADER_WEBGPU needing to be disabled when
WEBGPU=0.

Since we're changing some of the runtime flags in
/etc/chromium.d/gpu-options, we might as we make
some further changes.

Add some new (optional/commented) skia-related
flags to the WebGPU section :-

--enable-skia-graphite
--skia-graphite-backend

The first seems to be needed to enable the new skia
graphite feature. The second allows selecting the
backend for this (choices are: dawn-swiftshader
dawn-vulkan, dawn-metal and metal).

Also ensure that all flags in /etc/chromium.d/gpu
are commented, allowing users to choose which ones
to enable.