configure.sh: Fix disabling extensions in root menu
Disabling the previous incarnation was in fact
broken, but recent changes in effect disable
the patch for the new UI layout.

We rename the variable from EXT_TOOLS_MENU to
EXTENSIONS_ROOT_MENU, and swap the logic round.

In other words, we are disabling extensions in
the root menu, as opposed to enabling extensions
in the 'More tools' menu.