configure.sh: Fix disabling extensions in root menu
The previous incarnation was in fact broken if
one was trying to keep extensions in the root
menu. I doubt many noticed, as I doubt many
people are fans of chromium cluttering up
the menu.

As well as updating the patch due to upstream
changes, the EXT_TOOLS_MENU variable is renamed
to EXTENSIONS_ROOT_MENU, and the logic inverted.
This should be a bit more intuitive.