Yet another change to the vaapi flags
In the ever-changing universe of chromium VA-API
acceleration, this might work for some, but
inexplicably not for others.

I have removed a few seemingly obsolete options
and included an experimental nvidia flag.

I have also linked to the official documentation
(even this admits that VA-API on linux is not
officially supported.