Add -Wl,--lto-CGO3 to ldflags when using -O3
This adds '-Wl,--lto-CGO3' to ldflags within the
'optimize_speed' config (this config is used by
the build system to optimize at the -O3 level).

This linker option raises the lto codegen level
from 'default' to 'aggressive' to give a bit more
optimisation to the linker for code when the -O3
level is in effect.

Only parts of V8 are compiled at the -O3 level.
Interestingly, these use an lto optimisation
level of zero (--lto-O0), albeit with default
codegen optimisation.

We just turn up the lto codegen optimisation to
match that clang.

More info here: https://reviews.llvm.org/D141970