Enable google labs via flag file switch
Instead of disabling a patch in the configuration script
when LABS_TOOLBAR_BUTTON is enabled, we enable a switch
in /etc/chromium.d/ui. Users can now elect to enable
the button via a simple edit to the flag file.

We can now apply the patch regardless of configuration :-

labs-toolbar-button.patch is moved from optional/disable
to various/disable.

Also note that not enabling the button prevents the
corresponding entry in the 'More tools' submenu from
being visible. I had previously assumed that the menu
entry would remain.