Update Toolchain.md to now strip installed binaries
I was perhaps a bit too cautious in recommending not
to strip the clang/llvm binaries when installing them.

There was an issue with stripping which caused segfaults
on stripped bolt-optimised binaries, but upstream did
actually introduce the '--use-gnu-stack' to llvm-bolt
to work around this issue.

So change to using the ninja install/strip command.
Also tweak the logic on a couple of stat commands so
that we check for 0755 permissions instead of 0700.