Enable hls demuxer by default
This was really a work in progress, because the
demuxer without the player component isn't really
of much use.

Things have moved on now with new commits merged
recently, so let's try enabling the demuxer to be
built by default.

The new hls-player flag file contains a couple of
flags which need to be uncommented for the player
to work. These are :-

--enable-builtin-hls
--enable-features=HlsPlayer

We change the name of the HLS_DEMUXER configuration
variable to HLS_PLAYER, and use the values below :-

0 - disable
1 - enable
2 - fully enable (with the flags unhidden)

The default is 1.