Add --disable-grease-tls and --http-accept-header
The '--disable-tls-grease' flag in the network flag
file seems to be an orphan from an old removed patch.

There's a differently named '--disable-grease-tls'
flag that was upstreamed, so let's add that instead.

Disabling TLS grease is mainly of interest to those
wanting to use UC with tor. For this it needs to be
combined with --http-accept-header [1].

The grease flag could be used on its own, but bear
in mind the point of the feature is to prevent
problems due to misconfigured servers.

[1] See upstream UC issue #783