Re-work DNS configuration
There are two points to make here.

Firstly, DNS_BUILTIN should not depend on the
patch removing the dns config service.

Secondly, DNS interception detection appears
to depend upon this service.

A new DNS_CONFIG variable is introduced to
control the enabling/disabling of this service.
We exit with an error if both DNS_INTERCEPT=1
and DNS_CONFIG=0.

Also document the basics of managed policy and
enabling the builtin DoH client.