Set WebRtc IP handling policy via managed policy
The reason for dropping the ability to set this
via the flag in /etc/chromium.d/network is that
for me this didn't actually work.

I had tested this in the past with ip addresses
not being leaked, but somewhere along the line
it stopped working.

Setting the flag via chrome://flags works as
expected, but to ensure that things work out
of the box we'll need to set it via managed
policy.

We set the 'WebRtcIPHandling' key to a string
value of 'disable_non_proxied_udp' in the
policy file.

Also document this in README.md.