Error out when NON_FREE=0 and SYS_OPENH264=0
Instead of making assumptions about what a
user wants, give them the chance to correct
misconfigurations themselves.

In this instance, a user might want to change
NON_FREE to 1 instead of having SYS_OPENH264
forced to 1.

This way, no files get touched and the script
can simply be re-run instead of having to
delete and start over.