Make POLLY=0 the default
We now error out if SYS_CLANG=0 and POLLY=1
instead of silently setting POLLY=0 (note
that SYS_CLANG=0 is the default).

We need to change the default for POLLY
otherwise the script errors out if run
without setting any variables.

Arguably POLLY is a slightly obscure/unsafe
optimisation, so this is perhaps another
reason to not enable it by default.

Also update configure-variables.txt which
got neglected when the previous polly
variables were amalgamated into one.