Depend on libgtk-3-0t64 for unstable builds
Let's follow the debian chromium package in
depending on libgtk-3-0t64. Debian stable is
not part of the transition, so in this case
we use the configuration script to switch
the dependency back to libgtk-3-0.

Also tag DEB_BUILD_MAINT_OPTIONS=abi=+time64
onto our build command for unstable releases.