Disable enterprise watermark
This was done to work around one of those
hard-to-solve undefined symbol errors (where
the header containing the symbol is present
but somehow is not visible to the linker).

Enterprise watermark is a new feature, and a
patch is needed to fix build errors arising
from compiling with enterprise_watermark=false.

Added: various/disable/enterprise-watermark.patch