Disable visual query by default
Most users probably don't ever use visual
search (or even the side panel which
houses this facility).

With potential privacy concerns over the
visual searching of webpages to generate
search suggestions, disabling this by
default seems to me to be reasonable.

Also update the list of components in
README.md that can disabled.