configure.sh: Fix chromedriver script re-run issue
This affects builds configured with DRIVER=0.

We can normally re-run safely after errors if
the error occurs before any files are touched.

The problem here was that no other files apart
from the chromefriver ones were touched before
hitting an error. The script cannot be re-run
because it errors out when trying to delete
these files a second time.

Avoid this situation entirely by moving the
file deletion to near the end of the script.