d/rules: Use a static list to restore bundled backup files
The files that are replaced by the unbundling process
can be found in build/linux/unbundle/replace_gn_files.py.
We are therefore able to produce a list of files based
on libraries which are potentially unbundled.

The list is kept in debian/misc_files/bundled.txt and we
use it to loop through each file (note that we add a
'.bundled' suffix) and do the restoration if it exists.

This should be a fair bit quicker than using a find
command, especially on rotational disks when the file
cache is cold. Note that the same restoration process
is run at the start and end of a build.