configure.sh: Fix dollar sign sed error
The eval command doesn't seem to process the
dollar sign in a couple of sed commands. This
causes unterminated 's' command errors.

Escaping the dollar with a backslash doesn't
seem to fix things, so use escaped brackets
to copy the whole line and append the new
text onto the end of it.