configure.sh: Fix and improve unbundling
There are a ton of changes in this commit.

The dependency chains involve :-

fontconfig, brotli, libpng, libxslt,
libxml, icu, harfbuzz-ng and freetype

We can pivot what gets enabled or disabled around
brotli and icu, because these seem to be the ones
which aren't available or require a trip to the
debian packages wayback machine.

So let's replace SYS_FREETYPE and SYS_HARFBUZZ
with SYS_BROTLI, and work out the permutations
of how to change things based on whether or not
we unbundle brotli or icu.

Other changes are listed below :-

Stopped enabling harfbuzz by default due to a
dependency on icu. The latter is disabled by
default due version 73 being uninstallable via
a package manager.

Treat SYS_ICU as being disabled by default and
comment things out in rules.in and control.in.

Remove incorrect info "dav1d depends on libavif"
from d/rules.in.

Add libpng-dev dependency to d/control.in for
when both SYS_BROTLI-0 and SYS_ICU=0.

Exit with an error message if both SYS_ICU=1
and STABLE=1

The use_system_libtiff=true flag is only needed
by pdfium xfa support. So enable when PDF_JS=1
and alse enable libzstd unbundling when this
is the case.

It's also no longer worth mentioning libtiff
being a dependency of libwebp. It seems best
to treat them as independent.