Drop libwebp
The webp, avif and aom libraries all seem to be
inter-related, and in general a dependency hell.

Dealing with unbundling dependencies is one of
the most time consuming and confusing things to
get involved in, and very easy to get wrong.

Trying to fix all this stuff mid-release is not
a good idea, although it did force me to have
a good look at a few things.

Unbundling of stuff like webp is more trouble
than it's worth, so lets drop it.

Dropped :-

optional/system/unstable/libwebp-sharpyuv.patch