Split and move skia bundled freetype/harfbuzz patches
Although the previous patch worked fine and
handled both bundled and system freetype and
harfbuzz, we can tune this and only patch
what we need to.

Bundled freetype is non-default, and bundled
harfbuzz is the default.

So have the freetype component of the skia
patch as a separate (commented-out) optional
patch.

Make the corresponding harfbuzz patch optional
too and have it enabled by default.

Disable the freetype patch when SYS_BROTLI=0
and enable the harfbuzz one when SYS_ICU=1.