Fix rare case of opentype-svg fix not getting enabled
To be absolutely correct we really ought to enable
fixes/opentype-svg-on-bundled-freetype.patch for
unstable, as long as SYS_BROTLI=0.

We cannot do this if patch enablement is only within
the STABLE=1 scope, so move it down to the SYS_BROTLI
scope below.

Since SYS_BROTLI=1 is the default for unstable, this
small bug wouldn't have affected many people.