configure.sh: Dynamically handle rustc version within d/rules
This is to make things more robust, prompted by
running into an issue myself.

The problem is with upgrading rust in between
a stop (or a build error) and a restart. If
one doesn't update the rustc_version build
flag (in d/rules)  with the new version, then
build errors can ensue.

Dynamically set the version within d/rules so
that the new rust version is automatically
determined and the build flag set accordingly.

This simplifies the configuration script, but
doesn't add any significant complexity to the
rules file.

Not having to execute rust or clang (on a
cold start) makes the configuration script
run at lightning speed once again.