configure.sh: Dynamically handle clang version within d/rules
As with rust, setting the clang version via d/rules
seems to be an overall simplification, as well as
having the nice side effect of speeding up the
first run of the configuration script.

The following internal variables have been changed :-

Previous instances of CLANG_VER have been replaced
by LLVM_VER. (LLVM_VER is not to be confused with
LLVM_VERSION, which is also in d/rules).

CLANG_VER has been re-purposed to replace the C_VER
configuration variable, in order to try and make it
more easily distinguished from the CC_VER and CR_VER
internal variables.

We rename CLANG_DIR to LLVM_BASE_DIR (not to be
confused with the special LLVM_DIR environment
variable, which is exported within d/rules).

LLVM_BASE_DIR is used in d/rules to both set the
value of the clang_base_path build flag, and to
assist in getting the path to the clang binary.

Using the LLVM prefix for these variables helps
them stand apart from the 'CL' and 'CLANG' ones
which are used for different purposes.

Note that none of the above changes affect bundled
llvm/clang (SYS_CLANG=0).

Also update the documentation around CLANG_VER.