configure.sh: Make PGO=1 fail with old clang versions
This turns the warning when setting CLANG_VER to
a version that's older than the one in d/rules.in
into an error.

We add a PGO-1 conditional because erroring out
when PGO=0 would of course be incorrect.

Also bump the llvm/clang version in the rules
and control files to that of the bundled
toolchain (currently 19).