configure.sh: Don't check for existence of control or rules
These existence checks were done near the end of
the script as a failsafe against the possibility
that the files went through unchanged and weren't
subjected to any sed edits.

We don't need to check rules or control, as the
former will always be touched by the rust/clang
paths/versioning. The latter will always be
altered via the AUTHOR variable.

Currently, only the INS and POL variables are
tested to see if they are empty (ie no edits to
make). So we only need to check these two.