configure.sh: Merge first SYS_CLANG section into the second
The setting of CR_VER can go under SYS_CLANG > 1
because we are using it as a source of truth as
to what the bundled version is.

CLANG_VER is meant to be used to change the
version in rules or control, and is therefore
put under SYS_CLANG=1.

The PGO compatibility can go under SYS_CLANG=1
for the moment, although could in future also
be used for SYS_CLANG=2.

Rmove the sanity check that makes sure that the
clang version in rules and control are the same.
It just gets in the way isn't really needed.