configure.sh: Use CLANG_VER to bypass clang version check
CLANG_VER is now set for SYS_CLANG=2 in addition
to SYS_CLANG=1. Additionally, when CLANG_VER is
specified, the clang binary version check is
bypassed in favour of a (faster) basic file
existence test.

If CLANG_VER is not set, then the script will
error out if you do not have a version of clang
installed which at least equals the value of
LLVM_VERSION in debian/rules.in.

The documentation has been updated to reflect
the above, plus it's made clear that the default
for CLANG_VER is supposed to be a major version.