configure.sh: Move the SYS_CLANG=1 block to the end
We can move LLVM_BASE_DIR out of the SYS_CLANG=1
block and put it with the other LLVM_BASE_DIR
assignment further up.

Then we can move the whole SYS_CLANG-1 block
to the end of the SYS_CLANG >=1 section.

This should make things easier to follow with
all the SYS_CLANG >= 1 and SYS_CLANG=1 stuff
more clearly demarcated.