configure.sh: Drop CC_VER and use CR_VER for version checks
When deciding whether to change the clang version
in d/control, we conpare CC_VER, the version in
d/control.in, to CLANG_VER, the override version.

By default the version in d/control.in is made
to match the version in d/rules.in (CR_VER).

So it makes sense to solely use CR_VER to see
whether the version needs to be changed in both
rules and control, as the versions in both
files should always be the same.

I just need to remember to update the control
file when bumping the clang version in rules.