configure.sh: Drop ability to disable supervised users
The disable supervised users patch seems to just
neuter the component as opposed to ensuring that
it's not built.

There doesn't seem to be any real advantage in
disabling this feature, and there's no binary
size saving.

Also the patch doesn't seem to work under linux
and causes a segfault.

It was nice to have when it worked, but longer
term is likely to break and be more trouble
than it's worth.